group reporter/requirements.txt
jieba_fast>=0.53
jinja2>=3.1.0
openai>=1.55.3
httpx>=0.27.0
playwright>=1.40.0
python-dotenv>=1.0.0
ijson>=3.2.0
